refactor: util 구조 변경 (#38)
